Move to shelf-stable hashing
- C# object hashing changes each time you launch a program! Who knew.
  Not me, thats for damn sure.
- Also did a few UI improvements.

+++ src/BSBackupSystem/Data/Migrations/AppDbContextModelSnapshot.cs
@@ -165,8 +165,8 @@ namespace BSBackupSystem.Data.Migrations
                         .HasColumnType("timestamp with time zone")
                         .HasColumnName("first_seen");
 
-                    b.Property<int>("FullHash")
-                        .HasColumnType("integer")
+                    b.Property<long>("FullHash")
+                        .HasColumnType("bigint")
                         .HasColumnName("full_hash");
 
                     b.Property<Guid?>("GameId")
@@ -177,8 +177,8 @@ namespace BSBackupSystem.Data.Migrations
                         .HasColumnType("timestamp with time zone")
                         .HasColumnName("last_seen");
 
-                    b.Property<int>("PreRetreatHash")
-                        .HasColumnType("integer")
+                    b.Property<long>("PreRetreatHash")
+                        .HasColumnType("bigint")
                         .HasColumnName("pre_retreat_hash");
 
                     b.Property<Guid?>("PreviousSetId")
